requirements.txt
pillow==12.1.1
zxing-cpp==3.0.0
